Add the "slurm topa" command
Whilst seeing the usage by user is useful it is also very useful for
us at VLSCI to see a summary of the usage by account (project) as well.

This change adds a "slurm topa" command to do just this, a very simple
change based on the existing "slurm top" code.